--project argument supports paths starting with '~'
It's handy to be able to be able to specify paths
starting with '~' rather than having to supply
the full path to a project. This commit calls
os.path.expanduser when building a Project from a path,
which is safe because if the path cannot be expanded
then it is returned unaltered.