Add widgets to new markdown editor for images and links
The default image and link widgets for the TUI markdown editor do not
integrate with the Lektor tools of uploading images, attachments, and other
pages. With the creation of these two custom widgets for adding images and
links, the editor is integrated into the Lektor editor more fully.

The image widget allows:
 - adding currently uploaded attachments
 - uploading images as attachments, and then adding them in
 - adding images from external url

The link widget allows:
 - searching for an existing page on the lektor site to link to
 - adding an external url as a link