Make flow widgets pass down the routing props
Previously the flow widget did not pass down the routing props to all widgets
containd within it, which broke the markdown-gui editor (as it requires them)
when used within a flow block. This commit makes the flow widget pass them
down, and also prevents basic html elements from trying to use them when they
shouldn't (such as the primative widgets).