make serialization of widget data occur once upon load
when data was loaded from the API, it was put into `recordInitialData`, and
it would be deserialized whenever needed. this caused unnecessary re-rendering
of flow widgets due to a issue with flow's `deserializeValue` function.

now when data is loaded from the API, it is all deserialized where necessary,
removing the requirement to do it on the fly multiple times.

this makes more sense, as data is now deserialized when it is loaded from the
API, and serialized when the page is saved.