make flow's `deserializeValue` a pure function
the `deserializeValue` function of flow modified the `lastBlockId` variable.
As this was used for creating the `key` props for the flow widgets, evey time
this was called, the flow widget would re-render needlessly.

this makes it a pure function, which combined with modifying the way the
`deserializeValue` function is called, stops the unnecessary re-rendering.