fix CI
- Node 8, 11 have reached EOL, Python 3.8 shouldn't be allowed to fail
- Add Node 13
- Fix some pylint warnings and ignore some

Also reduce CI matrix:

Python and Node are quite independent, so it's not necessary to test
them in all possible combinations. It's probably also a waste of the
free resources provided by Travis and Appveyor to run CI on all
supported Python versions (and not just the oldest and newest), so
remove some entries from the test matrix.