change test that breaks git configuration
This test was testing a one-line function that basically just called out
to `git config` to get the configured email with a test fixture that
had 3 calls to `git config` and would break the local git config.

Testing some functionality by having a test fixture that just duplicates
that code seems somewhat pointless, even more so when that breaks the
test runners git configuration.

If removed the fixture, instead we now only test that the function
returns a string.