Prevent fsync storm during update_all_source_infos
This applies the fix (for #144) that was applied in 701a37c8 to
Builder.build_all to Builder.update_all_source_infos.

On my old machine, on a site with about 100 sources, this change reduces
the time for a source info update from 17 to under 2 seconds.  (A source
info update is run at the beginning of each ``lektor server`` run.)