Implement slug styles (name TBD)
Slug styles specify how slugs are turned into file and directory names
in the URL path. This is roughly based off the design in
https://github.com/lektor/lektor/issues/809#issuecomment-687136459.

As is, this PR fixes #344 and deals with the tricky "global" part of
doing #809 properly. It also enables pages with dots in their slugs to
be paginated and fixes some issues with the dev server resolving
descendants of such pages incorrectly. It is still a prototype, though
(I have only done some casual manual testing and yet to write any
tests).

In addition to a sitewide default setting and per-page setting, we would
likely also want datamodels to be able to specify the slug style of
their pages and/or their child pages. To stay similar to the current way
`slug_format` can be configured, it seems we should also allow
datamodels to specify the slug styles of their children, but as
mentioend in #806 that behavior is a little strange and worth
reconsidering, so this PR doesn't try to implement any such behavior.