WIP: deprecate .source_filename, cleanup .iter_source_filenames
When alts are enabled, Page.source_file was being generated based on
the page.alt.  This is not correct, since the alt-specific file may
not exists.  E.g. if there is only a `contents.lr`, for `alt="en"`
`source_file` was returning the non-existent file `contents+en.lr`.

In general, `.source_filename` and `.iter_source_filenames` should
return only extant source filenames.

Dependency tracking for the blog pages of the example site is not
working.  They get rebuilt every build.  This has something to do with
the `Siblings` virtual source object for the blog pages not listing
the correct source filenames due to the issue with Page.source_file
described above.