Increase timeout in `IterateInThread` test helper
The github CI workflow tests seem to occasional fail do to a timeout
in `IterateInThread.__next__` while waiting for the other thread to
produce an item.

Here we increase the timeout from 0.1 to 2.0 seconds in the hopes of
preventing that failure.

Since the timeout is never hit except upon test failure, this should
not result in an increase in running time for any (passing) tests.