Refactor: Move adaptor logic from lektor.admin.modules.api to Tree*
The `Tree`, `TreeItem`, and `Alt` classes exists to serve as an
adaptor from the db to a view that serves the needs of the API.

Over time, a bunch of adapter logic has crept into the API code —
things like the sort-order children, etc.  This moves some of it back
to the Tree/TreeItem code.