MultiLineTextInputWidget: use hooks; remove scroll-adjustment
The scroll-adjusting seems to cause various bugs in different browsers.
Sticking to the default of having the browser scroll if the textarea
changes seems to work better.

The auto-sizing still works by setting the height to 'auto' and then to
the scrollHeight directly afterwards. This is also the current top
answer on the stackoverflow question on auto-sizing textareas.