Clean up MANIFEST.in
Potentially, in dev environment, a lot of cruft was being included in
dists. This results in overly large installs in tox envs.

This commit tightens up the recursive-include rules in `MANIFEST.in`.

It also adds [check-manifest][] to the the commands run in the `lint`
tox env, in an attempt to make sure we're not leaving anything out.

[check-manifest]: https://pypi.org/project/check-manifest/