Clean up SourceObject.url_to and implement "lektor:" URL scheme
This add a "lektor:" URL scheme, recognized by `SourceObject.url_to`,
which causes the URL path to be interpreted as a (possibly relative)
Lektor db-path.  This is very similar to how plain paths (with no
scheme or netloc) are handled currently, except that, if resolution
to a Lektor source object fails, an error will be raised, instead of
falling back to interpreting the path as a URL path.

It also introduces a slightly-backward-incompatible change in how
`SourceObject.url_to` interprets its `absolute` parameter.
Formerly, if `absolute` was truish, the "absolute" URL path relative
to the site root (configured by the `[project]url` setting in the
project file). This has now been changed to return an absolute URL
path (which includes any prefix defined by `[project]url`.)

It can be argued that the former behavior was a bug.  Suppose a
template, not unreasonably, contains

    See <a href="{{ this.get('child').url_to(absolute=True)
    }}">child</a>.

That link will be a broken one if `[project]url` is set to something
with a non-empty path (e.g. `url = http://example.org/site-prefix/`).