Fix installation of ImageMagick on Windows test runner
We had been installing the `imagemagick.tool` Chocolatey package.
That package seems to be neglected and has not been successfully
installing for a few months now.

We switch here to installing the `imagemagick.app` package instead.
It installs successfully, but, for whatever reason it installs to a
directory under `\Program Files`, and Chocolatey is unsuccessful at
it's attempts to add the install directory to `$PATH`. So we have to
manually figure out where ImageMagick was installed to and add that
to the search path.