fix(api): /matchurl should only match URLs for Records
This is a backport of 0598aac from the master branch.

Since /matchurl is used by the admin frontend to track which record to
edit when the preview iframe is navigated (and since the admin
frontend only knows how to deal with Records — i.e. pages and
attachments) this method should only match URLs that correspond to
Records.

This also fixes a exception that occurs when /matchurl is
called with a URL to a VirtualSourceObject.