The Siblings virtual source does not depend on siblings' content.
The content of the Siblings virtual source object only depends on
the identity of the siblings, not on the siblings' content.  I.e.
the Siblings object should only be considered changed when the
identity of one or both of the siblings is changed.

(Any artifact which in fact uses data from one of the siblings
will also register a dependency on that sibling which will trigger
a rebuild if the sibling is touched.)