Make the identity key (source, alt) explicit for VirtualSourceObjects
Both Records and VirtualSourceObjects are described by the identity
key (path, alt, pad).  That is, if

   obj2 = pad.get(obj.path, alt=obj.path)

then

   obj2 == obj

Records already had __eq__ and __hash__ methods which made this
explicit.  Here we add those methods for VirtualSourceObjects.

Note this allows us to convert Context.referenced_virtual_dependencies
from a dict to a set.  (The key was only being used as an identity key
for the VirtualSourceObjects.)