Fix Lektor installation of extension modules required by plugins.
The lektor-website/packages/project-data plugin requires
`readme_renderer[md]`, which, in turn requires `cmarkgfm<0.7.0`.
There is no wheel appropriate for the Windows GitHub runner
for `cmarkgfm==0.6.0`, so when Lektor goes to install the plugin,
the extension must be compiled.  Apparently the `ProgramData`
and `ProgramFiles` environment variables are important to this
process.