admin: use esbuild as a bundler for the frontend
From my experience with it in recent months, esbuild is vastly superior
to webpack (and all other bundlers). We are using modern TS and React
which are supported out-of-the-box by esbuild, allowing the rather
opaque build tools webpack and babel. Only for SASS we need a plugin,
which is trivially done in a couple of lines of code.

Reduces the build time from ~12s to ~3s on my machine (even though it is
adding a minification step which we previously did not have).