Add missing implict requires, remove dropped extras & unneeded deps
Add MarkupSafe and pytz as direct runtime deps, since they
are imported and used but not explictly declared, and are only
satisfied transitively which cannot be relied upon.
Also, remove the [security] extra to requests as it was dropped
upstream, as modern, more security SSL is included with Py>=3.6-3.7,
which Lektor already requires, and the [toml] extra to Setuptools_SCM,
which was removed in 6.2.0.
Finally, remove wheel as a direct build-requires following current
normative guidance, as it will be pulled in as needed by Setuptools.