admin: handle hotkey events delivered to preview iframe
When the preview iframe has the keyboard focus, key events do not
bubble up to the parent window.  Here we attempt to pass a copy of
all keydown events delivered to the iframe on to the top-level window.

This solves at least part of the "Ctl-e doesn't work" problem reported
in #1022.