CSS-based solution to autosize textarea
This (I hope) fixes the wonky scrolling described in #1038.

This is a backport of the first commit in PR #1050.  It omits all
the hairy bits required to enable manual drag-handle resizing of
textareas.

(I want to get this into the stable 3.3 branch, so that we can have
a working release.)

Based on ideas from this post:

https://css-tricks.com/the-cleanest-trick-for-autogrowing-textareas/