An alternative fix for installing local plugins.
Here we create a "nested" virtual environment in which to install our
plugins.  By "nested" we mean that it keeps the `site-packages`
directory of the "containing" virtual environment (or the system, if
there is no containing virtual environment on sys.path. (It does this
via some `.pth` file magic.)

This allows us to use plain old `pip install` to install plugins,
rather than having to resort to `pip install --target <dir>` which
was causing all sorts of grief.

This also allows us to generalize our plugin installation machinery to
support any PEP517 / PEP660 compatible distribution.

We've also cleaned up detection of changed requirements.

Previously, the package cache was created whenever:

- Any new plugin (local or remote) was added to the project
- Any plugin (local or remote) was removed from the project
- The specified version for any remote plugin was changed

Now that we are using a real virtualenv to manage installation, we
could probably loosen these requirements, and let pip deal with, e.g,
reconciling changes in requested package version.

For now, we keep the existing behavior, but clean up the logic for
detecting when requirements have changed.