feat: tool drawer for edit and live-reload buttons
This replaces the *edit pencil* on HTML artifacts served by the admin
server with "tool drawer".

The tool drawer contains an edit button (which takes the place of the
edit pencil), as well as a toggle button that can be used to disable
live-reload (see #1027) for the window.

If the tool drawer is obscuring part of the HTML page that one would
like to look at or click on, the drawer may be dragged vertically
to reposition it.  It may also be (mostly) hidden off screen to
minimize its visual impact.