refactor: remove dependency on pytz
Since the introduction of PEP 614 (the `zoneinfo` module in the
standard library, available since python 3.9), the future is to
transition away from `pytz` to `zoneinfo` for timezone support.

Babel, since 2.12, will now work with either `pytz` or `zoneinfo`,
and no longer requires `pytz` on pythons >= 3.9.

This change also fixes some hokey-ness in the previous implementation
of DateTimeType.