refactor(watcher): replace Queue with flag (threading.Event)
The current implementation stores all unprocessed filesystem change
events in an unbounded Queue.  There can possibly be quite a number of
events, so this can be (in some situations) a memory consumption
issue.

All we really care is whether there has been *any* interesting event
since the last time we checked. All we need for that is a flag, not a
queue.