fix(tests): increase wait for stale events on macOS
Under macOS, the FSEventObserver (at least) seems to return events for
filesystem changes that happened a bit for the Observer was started.
We had a short wait in place to let these stale events go by. It
mostly worked, but there have been sporadic spurious failures on the
macOS GitHub test runner.  Here we increase the wait time, but — so as
not to slow tests on other platforms — only perform the wait when
running on macOS.