Fix for spurious rebuilds with watchdog>=2.3.0 (#1117)
* fix(watcher): ignore FileOpenedEvents

Since watchdog==2.3.0, its InotifyObserver emits events whenever a
file is opened. This was causing spurious Lektor rebuilds.

* refactor(watcher): optimization: we only care about file events

* fix(tests): increase wait for stale events on macOS

Under macOS, the FSEventObserver (at least) seems to return events for
filesystem changes that happened a bit for the Observer was started.
We had a short wait in place to let these stale events go by. It
mostly worked, but there have been sporadic spurious failures on the
macOS GitHub test runner.  Here we increase the wait time, but — so as
not to slow tests on other platforms — only perform the wait when
running on macOS.