Refactor(watcher): replace Queue with Semaphore (#1118)
* refactor(watcher): replace Queue with flag (threading.BoundedSemaphore)

The current implementation stores all unprocessed filesystem change
events in an unbounded Queue.  There can possibly be quite a number of
events, so this can be (in some situations) a memory consumption
issue.

All we really care is whether there has been *any* interesting event
since the last time we checked. All we need for that is a flag, not a
queue.

* refactor(watcher): skip is_interesting check if event already set

* refactor(watcher): remove unused params from is_interesting

Remove the unused `time` and `event_type` parameters from
`Watcher.is_interesting()`.

* refactor(watcher): delete lektor.watcher.watch