fix(packages): fix local plugin installation
The most recent version of pip (23.1) removes support for the
`--install-option` parameter to `pip install`.  This breaks our
`install_local_package` function, which uses `--install-option`
because `pip install --target` didn't used to work with `pip install
--editable`.

As it turns out, recent versions of pip support the use of `--target`
with `--editable`.  So here we make use of that.

Note that this also relieves us of the need to parse our local
plugin requirements ourselves (thus, hopefully, fixing #865.)