test: do not run certain slow tests in every testenv variant
Here we mark a few particularly slow tests, and edit tox.ini so that
these tests are not run (by default) in the "variant" testenvs
(e.g. py311-tzdata).

The tests marked with `slowtest` are particularly slow on the
macOS (and probably the Windows runners). Omitting them saves about
a minute per testenv in those environments.