refactor(imagetools)!: name thumbnail files based on final parameters
Previously, thumbnail artifact names were generated by adding suffix
that serializes the parameters passed to make_image_thumbnail.

Especially when using "fit" mode, there are many different sets
of parameters to make_image_thumbnail that can generate the same
thumbnail.

Here we change the method of generating the thumbnail suffix
so that it serializes the parameters to ThumbnailBuilder.
These include the full dimension of the thumbnail (both width and
height).

This can prevent the generation of multiple identical thumbnails
(with different names). I think the resulting code is cleaner, as well.