fix(frontend): address "save & preview" race condition
The current code marks the edit form as clean after a successful
save request. However, if the form was further edited after the
save request was sent, but before the response was recieved, this
is wrong. (As a result, edits can be lost during a "save & preview".)

This fixes that.