fix: do not override user-specified warnings filter
This is a backport of e2d0274.

Currently the Lektor CLI unconditionally sets the warnings filter to
"default". There is no way of the user to configure this.

Here we bypass re-configuring the warnings filter if it has already
been customized, either via the PYTHONWARNINGS environment variable or
by python's -W flag.

E.g., now one can show all warnings by doing either of:

    PYTHONWARNINGS=always lektor build

or

    python -W always -m lektor build

Ref: https://docs.python.org/3/library/warnings.html#overriding-the-default-filter