fix: pin werkzeug<2.4, ignore deprecation warnings in tests
Our Publisher API involves passing a `werkzeug.urls.URL` instance to
`Publisher.publish`.  As this is an API assumed by custom publishers
implemented by Lektor plugins out in the wild, we can not simply
change the API.

PR #1143 — in the master branch — implements one approach to updating
the API in a backward-compatible manner, but it is rather involved.

For the sake of the 3.3.x branch, for now we pin werkzeug<2.4 and
leave the Publisher API as it is.