fix(builder): logic in FileInfo.unchanged was incorrect
This is a longstand bug — there was a race condition.

For efficiency, Lektor's build system caches filesystem os.stat() results and
(sometimes) checksums of file content.  (For this is uses PathCache to
cache FileInfos which, in turn, cache stat and checksums for
individual files.)

The FileInfo classes compute defer computation of stat results and
checksums until they are first accessed.

`FileInfo.unchanged` is a method that is used to compare a historical
FileInfo which has been retrieved from the build database with
"current" state of a file.  The logic in `FileInfo.unchanged` was
subtly flawed.

It first compared the `size` and `mtime` of the file. For
regular (i.e. non-directory) files it returned `True` early if those
matched. If they differed it went on to compare checksums, and
returned a result based (only) as the checksum comparison.

The logic is flawed: if the size of the file has changed, we can be
assured the file has indeed changed, and should return early right
then — no need to check the checksum.

The real bug arises due to the fact that values of size/mtime and the
checksum that are stored in the build database may have be computed
and cached at different times. It's possible, in the case that a file
is modified during a build cycle, for the size/mtime stored in the
build database to correspond to it's pre-modification state, while the
checksum corresponds to its *post*-modification state.  In this case,
due to to above mentioned logic flaw, a file will be detected as
unchanged, even though its size has changed.

There is a comment in the code that indicates the intent was to skip
computing and checking checksums for regular files if their size and
mtime match.  Here, we fix the logic to do just that.  Another option
would be to check checksums as well (but only in the case that the
size matches.)

Going forward, we probably should take care to ensure that whatever
we record about source file state in FileInfo and the build database
represents the file state at a single point in time.