refactor: capture FileInfo state early
In order to minimize the possibility of race conditions, the idea here
is to capture source state (mtime, size, and possibly checksum) as
early as possible in the build process.  This allows us to detect
(and rebuild during the next build cycle) the case where a source file
changes during the time that an artifact is being generated.

Previously the computation/caching of those file metadata was deferred
until such time as they were requested. (That deferral did not result
in any saving of computational effort, since, at the end of the build
the metadata is always computed and written into the buildstate
database.)