refactor(lektor.assets): present a single merged asset tree
When themes are in use, assets can be located in one of multiple asset
directories.

Previously, in this case, we dealt with multiple asset roots, each
with their own heirarchy of descendant assets. Some of these
descendants can be shadowed by assets in one of the other asset roots,
and should be ignored. This is difficult to manage.

Here we refactor the code in `lektor.assets` to present a single
logical merged asset tree which overlays the assets from the primary
project `assets` directory with the `assets` directories of any
themes.