Fix #908: Omit shadowed assets from build (#1147)
* test: add test to exercise #908

* refactor(lektor.assets): present a single merged asset tree

When themes are in use, assets can be located in one of multiple asset
directories.

Previously, in this case, we dealt with multiple asset roots, each
with their own heirarchy of descendant assets. Some of these
descendants can be shadowed by assets in one of the other asset roots,
and should be ignored. This is difficult to manage.

Here we refactor the code in `lektor.assets` to present a single
logical merged asset tree which overlays the assets from the primary
project `assets` directory with the `assets` directories of any
themes.

* refactor(lektor.assets): be more careful about validating path components

* refactor(lektor.assets): refactor Asset.resolve_url_path

This cleans up the logic for detecting cases where the URL (or asset)
name differs from the source file name. (There are cases where the
suffixes of the names can differ.)

This disuses and deprecates the `from_url` parameter to
`Asset.get_child`.

* refactor(lektor.assets)!: remove vestigial method `Asset.build_asset`

`Asset.build_asset` is an undocumented no-op. It appears to be
unreferenced by Lektor itself as well as all Lektor plugins that I've
been able to find.

* tests(lektor.assets): enough annotations to get mypy to pass

* tests(lektor.assets): increase test coverage

* fix: do not Path.resolve() asset roots

This was breaking the (imperfect) logic in
`lektor.builder.PathCache.to_source_file` which checks that source
files are contained within the project tree.

* refactor(lektor.assets): cache child assets on Directory

Optimization.

Since Assets are tied to a Pad, the cache lifetime will match that of
the Pad. This seems appropriate.