fix: `bag()` should get `site` from jinja context
Jinja global functions should obtain `site` from the jinja context
rather than from the Lektor build Context.

Sometimes `site` is explicitly set in the jinja context. In that case,
that is the value that should be used.  Furthermore, outside of an
artifact build, a Lektor build Context is not available.