fix(bitrot): use PIL.Image.Transpose int enum
Pillow >= 9.1 moves many of the constants that were in the top-level
of the PIL.Image module into IntEnums within the module.  But then it
re-exports the enum members to the module programmatically.
Something has changed recently (I'm not sure whether on the Pillow
side or the pylint side) to make pylint throw `no-member` warnings
when these are accessed as module-level attributes.

For the sake of moving forward, here we use the enums, polyfilling
them if running with an older version of Pillow.