feat(admin gui): use link for "Return to Website" button
Using a real link (`<a href="...">`) instead of a button with onclick
handler allows the user to, e.g., open a new window with a
live-preview by middle or right clicking on the button.

Anyway, since the "button" leads to a new URL outside of the admin
app, it's the right thing to do.  E.g. see
https://css-tricks.com/buttons-vs-links/.