Refactor: Use Pillow for image info (#1138)
* refactor(imagetools): refactor get_image_info to use Pillow

Use Pillow to determine image format and dimensions rather than our
own home-rolled code.

* fix: support Pillow<7

* refactor(imagetools): use Pillow to access EXIF tags

* fix: compatibility with older versions of Pillow

* tests(imagetools): type annotations

* fix: do not transpose PNGs based on Exif Orientation

* refactor(imagetools): deprecate passing file object to get_image_info

* refactor(Video): use cached_property

* fix(imagetools): always convert image to RGB before scaling

* refactor(imagetools)!: remove deprecated methods of ThumbnailMode

Remove deprecated property `ThumbnailMode.label` and classmethod
`ThumbnailMode.from_label`.

* refactor: split imagetools from module to package

* fix(b/c): re-export `compute_dimensions` and `get_quality`

The following plugins appears to use internal bits of
`lektor.imagetools`:

- https://github.com/numericube/lektor-minification/
- https://github.com/chaos-bodensee/lektor-image-resize/
- https://github.com/skorokithakis/lektor-thumbnail-generator/

From a perusal of their code, it appears that they'll all be
broken by recent Lektor code changes having to do with moving
from ImageMagick to Pillow.

This doesn't completely address the issues for those plugins,
but, for the sake of backward-compatibility, here we:

- re-export `lektor.imagetools.compute_dimensions` (which was moved to
  `lektor.imagetools.thumbnail`).

- re-implement and deprecated `lektor.imagetools.get_quality`

* docs(imagetools): update comments

* tests(imagetools): use real Context in tests

Instead of a Mock, use a real lektor.context.Context in the image
tests.

* test(imagetools): add thumbnail dependency race-condition test

* fix(test): fix test for pillow 10.0.0

I'm not sure why this test ever worked.  Different versions of convert
from color strings to HSV values differently.

* refactor(imagetools): clean up get_svg_info

- rename to `_get_svg_info` to emphasize it is internal
- raise exception rather than returning "unknown" if SVG is unparseable
- rewind file after _get_svg_info call

* fix(bitrot): use PIL.Image.Transpose int enum

Pillow >= 9.1 moves many of the constants that were in the top-level
of the PIL.Image module into IntEnums within the module.  But then it
re-exports the enum members to the module programmatically.
Something has changed recently (I'm not sure whether on the Pillow
side or the pylint side) to make pylint throw `no-member` warnings
when these are accessed as module-level attributes.

For the sake of moving forward, here we use the enums, polyfilling
them if running with an older version of Pillow.

* chore: remove python<3.8 compatibility code