📦 Upload dists to PyPI in a dedicated job
This is important because it allows migrating to Trusted Publishing
safely [[1]]. Which in turn will enable the maintainers to revoke the
API token secret making the publishing tokenless and will start
uploading digital attestations.

The environment called `pypi` will have to be set up as trusted on the
PyPI side and might benefit from having manual approvals within the
repository settings.

[1]: https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/