🧪 Rename the CI/CD workflow to reflect contents
This patch renames the main GHA workflow definition file to mention
the CD bit (since it includes publishing to PyPI). It also changes its
name so that it does not take up a lot of space in the UI and is not
misleading in the same way. The job names remain the same, meaning
that well-known checks/statuses still appear in the PR UI.