🧪💅 Unify running linters via pre-commit
Previously, pylint was specialcased. However, there are a lot of
advantages to having a unified invocation. For example, a single
tox env run that can be integrated into many places. Or, enabling
pre-commit.ci eventually.

The only downside is having to keep track of a copy of the direct
runtime dependencies in another config.