fix(builder): stop treating @ as virtual by default
The bug report was about filenames starting with @ getting rebuilt every time.
Upon inspection, the path is treated like a virtual path.
I've changed the code so that the virtual information becomes part of
the database, thus the @ char in the path does not automatically mean
something special.

I did not find any migration scripts, so I did it directly in the
builder, dropping old tables.
This will force a full rebuild upon the first start.